Don't crash the backup when a DSE node has no metadata folder & Read from cassandra.yaml (#973)

+++ medusa/cassandra_utils.py
@@ -17,6 +17,7 @@
 
 
 import fileinput
+import functools
 import itertools
 import logging
 import os
@@ -266,6 +267,13 @@ class CassandraConfigReader(object):
         return pathlib.Path(saved_caches_directory)
 
     @property
+    def dse_metadata_directory(self):
+        # DSE adds this key to cassandra.yaml itself (not dse.yaml). Older DSE versions (eg 5.1, #826)
+        # don't have it at all, so callers must fall back to a guessed location when this is None.
+        metadata_directory = self._config.get('metadata_directory')
+        return pathlib.Path(metadata_directory) if metadata_directory else None
+
+    @property
     def listen_address(self):
         if 'listen_address' in self._config and self._config['listen_address']:
             return self._config['listen_address']
@@ -355,8 +363,15 @@ class Cassandra(object):
         config_reader = CassandraConfigReader(cassandra_config.config_file, release_version)
         self._cassandra_config_file = cassandra_config.config_file
         self._root = config_reader.root
-        self._dse_root = self._root.parent
-        self._dse_metadata_folder = 'metadata'
+        dse_metadata_directory = config_reader.dse_metadata_directory
+        if dse_metadata_directory is not None:
+            self._dse_root = dse_metadata_directory.parent
+            self._dse_metadata_folder = dse_metadata_directory.name
+        else:
+            # metadata_directory isn't set in cassandra.yaml (older DSE, eg 5.1, #826) - guess its
+            # location next to the data directory, which holds for the common/default DSE layout.
+            self._dse_root = self._root.parent
+            self._dse_metadata_folder = 'metadata'
         self._commitlog_path = config_reader.commitlog_directory
         self._saved_caches_path = config_reader.saved_caches_directory
         self._hostname = contact_point if contact_point is not None else config_reader.listen_address
@@ -438,9 +453,9 @@ class Cassandra(object):
                 session.execute(f"REBUILD SEARCH INDEX ON {fqtn}")
 
     @staticmethod
-    def _ignore_snapshots(folder, contents):
+    def _ignore_snapshots(metadata_folder, folder, contents):
         ignored = set()
-        if folder.endswith('metadata/snapshots'):
+        if folder.endswith(f'{metadata_folder}/snapshots'):
             logging.info(f'Ignoring {contents} in folder {folder}')
             for c in contents:
                 ignored.add(c)
@@ -463,9 +478,14 @@ class Cassandra(object):
 
         tag = "{}{}".format(self.SNAPSHOT_PREFIX, backup_name)
         if not self.dse_snapshot_exists(tag):
-            src_path = self._dse_root / self._dse_metadata_folder
-            dst_path = self._dse_root / self._dse_metadata_folder / 'snapshots' / tag
-            shutil.copytree(src_path, dst_path, ignore=Cassandra._ignore_snapshots)
+            src_path = self.dse_metadata_path
+            if src_path.is_dir():
+                dst_path = src_path / 'snapshots' / tag
+                ignore = functools.partial(Cassandra._ignore_snapshots, self._dse_metadata_folder)
+                shutil.copytree(src_path, dst_path, ignore=ignore)
+            else:
+                # Not every DSE node has a metadata folder - warn instead of failing the backup.
+                logging.warning(f'No DSE metadata folder found at {src_path}, nothing to snapshot')
 
         return Cassandra.DseSnapshot(self, tag)
 
@@ -493,7 +513,8 @@ class Cassandra(object):
         def delete(self):
             dse_folder = self._parent._dse_metadata_folder
             dse_folder_path = self._parent._dse_root / dse_folder / 'snapshots' / self._tag
-            shutil.rmtree(dse_folder_path)
+            if dse_folder_path.is_dir():
+                shutil.rmtree(dse_folder_path)
 
         def __repr__(self):
             return '{}<{}>'.format(self.__class__.__qualname__, self._tag)
@@ -576,8 +597,8 @@ class Cassandra(object):
         return False
 
     def dse_snapshot_exists(self, tag):
-        # dse files live one directory up from the data folder
-        # the root field should point to the data directory as defined in the cassandra.yaml
+        # _dse_root is metadata_directory's parent (from cassandra.yaml) when set, otherwise a guess
+        # based on the data directory - see Cassandra.__init__.
         for snapshot in self._dse_root.glob(self.DSE_SNAPSHOT_PATTERN.format('*')):
             if snapshot.is_dir() and snapshot.name == tag:
                 return True
